refactor(acl): unify key serialization between CLI and proxy
CLI (server.rs) now builds KeyEntry struct and serializes via serde
instead of manual JSON. save_keys_to_file in server_llm.rs also uses
serde_json::to_value(). Types moved to b00t-c0re-lib/src/acl_keys.rs.

Note: blocked by #537/#538 submodule issue (b00t-reflect-types missing)
